// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const UserRole = {
  USER: 'USER',
  STORE_ADMIN: 'STORE_ADMIN',
  SUPER_ADMIN: 'SUPER_ADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const TokenType = {
  EMAIL_VERIFICATION: 'EMAIL_VERIFICATION',
  PASSWORD_RESET: 'PASSWORD_RESET'
} as const

export type TokenType = (typeof TokenType)[keyof typeof TokenType]


export const StockJournalType = {
  IN: 'IN',
  OUT: 'OUT'
} as const

export type StockJournalType = (typeof StockJournalType)[keyof typeof StockJournalType]


export const DiscountType = {
  PRODUCT: 'PRODUCT',
  CART: 'CART',
  SHIPPING: 'SHIPPING',
  BUY_ONE_GET_ONE: 'BUY_ONE_GET_ONE'
} as const

export type DiscountType = (typeof DiscountType)[keyof typeof DiscountType]


export const DiscountValueType = {
  PERCENTAGE: 'PERCENTAGE',
  NOMINAL: 'NOMINAL'
} as const

export type DiscountValueType = (typeof DiscountValueType)[keyof typeof DiscountValueType]


export const VoucherType = {
  PRODUCT: 'PRODUCT',
  CART: 'CART',
  SHIPPING: 'SHIPPING'
} as const

export type VoucherType = (typeof VoucherType)[keyof typeof VoucherType]


export const OrderStatus = {
  PENDING_PAYMENT: 'PENDING_PAYMENT',
  AWAITING_PAYMENT_PROOF: 'AWAITING_PAYMENT_PROOF',
  PROCESSING: 'PROCESSING',
  SHIPPED: 'SHIPPED',
  DELIVERED: 'DELIVERED',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const PaymentMethod = {
  MANUAL_TRANSFER: 'MANUAL_TRANSFER',
  PAYMENT_GATEWAY: 'PAYMENT_GATEWAY',
  COD: 'COD'
} as const

export type PaymentMethod = (typeof PaymentMethod)[keyof typeof PaymentMethod]


export const PaymentStatus = {
  UNPAID: 'UNPAID',
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]